Conflicts:
	org.jage.student.examples-solutions/src/main/java/org/jage/examples/resource_meter/MonitoredIsolatedSimpleWorkplace.java

+++ org.jage.student.examples-solutions/src/main/java/org/jage/examples/resource_meter/ResourceMeterStrategy.java
@@ -1,7 +1,5 @@
 package org.jage.examples.resource_meter;
 
-import java.util.logging.Logger;
-
 import org.jage.property.ClassPropertyContainer;
 
 public class ResourceMeterStrategy extends ClassPropertyContainer implements
@@ -11,14 +9,14 @@ public class ResourceMeterStrategy extends ClassPropertyContainer implements
 	private MemoryInfo memoryInfo = new MemoryInfo();
 
 	@Override
-	public Long getCpuLoad() {
-		long cpuLoad = cpuInfo.getPercentageValue();
+	public Integer getCpuLoad() {
+		Integer cpuLoad = cpuInfo.getPercentageValue();
 		return cpuLoad;
 	}
 
 	@Override
-	public Long getMemoryLoad() {
-		long memoryLoad = memoryInfo.getPercentageValue();
+	public Integer getMemoryLoad() {
+		Integer memoryLoad = memoryInfo.getPercentageValue();
 		return memoryLoad;
 	}
